Add files via upload
The SQL query was written with many sub-queries to check each step. This has finally been cleaned up to use less sub-queries in place of more joins. Resources used reduced by 40% for 3-4 seconds it takes to run.

Added another query to create a transporting report, which details profit from buying in a city and selling in another, instead of crafting.